+++ pssh/clients/native/__init__.py
@@ -18,3 +18,4 @@
 # flake8: noqa: F401
 from .parallel import ParallelSSHClient
 from .single import SSHClient, logger
+from .sftp import SFTPClient